Refactor - extract ReportLogic class for easier testing
